tests: roll back metric/plots unification, introduce collect test
