+++ tests/func/metrics/test_diff.py
