+++ tests/func/plots/test_show.py
@@ -6,7 +6,9 @@ from collections import OrderedDict
 
 import pytest
 
+from dvc.exceptions import NoMetricsParsedError, OverlappingOutputPathsError
 from dvc.main import main
+from dvc.path_info import PathInfo
 from dvc.repo import Repo
 from dvc.repo.plots.data import (
     JSONPlotData,
@@ -21,8 +23,8 @@ from dvc.repo.plots.template import (
     TemplateNotFoundError,
 )
 from dvc.utils.fs import remove
-from dvc.utils.serialize import dump_yaml, dumps_yaml
-from tests.func.metrics.utils import _write_csv, _write_json
+from dvc.utils.serialize import dump_yaml, dumps_yaml, modify_yaml
+from tests.func.plots.utils import _write_csv, _write_json
 
 
 def test_plot_csv_one_column(tmp_dir, scm, dvc, run_copy_metrics):
@@ -704,3 +706,45 @@ def test_show_from_subdir(tmp_dir, dvc, caplog):
 
     assert f"file://{str(subdir)}" in caplog.text
     assert (subdir / "plots.html").exists()
+
+
+def test_show_malformed_plot(tmp_dir, scm, dvc, caplog):
+    tmp_dir.gen("plot.json", '[{"m":1]')
+
+    with pytest.raises(NoMetricsParsedError):
+        dvc.metrics.show(targets=["plot.json"])
+
+
+@pytest.mark.parametrize("clear_before_run", [True, False])
+def test_metrics_show_overlap(
+    tmp_dir, dvc, run_copy_metrics, clear_before_run
+):
+    data_dir = PathInfo("data")
+    (tmp_dir / data_dir).mkdir()
+
+    dump_yaml(data_dir / "m1_temp.yaml", {"a": {"b": {"c": 2, "d": 1}}})
+    run_copy_metrics(
+        str(data_dir / "m1_temp.yaml"),
+        str(data_dir / "m1.yaml"),
+        single_stage=False,
+        commit="add m1",
+        name="cp-m1",
+        plots=[str(data_dir / "m1.yaml")],
+    )
+    with modify_yaml("dvc.yaml") as d:
+        # trying to make an output overlaps error
+        d["stages"]["corrupted-stage"] = {
+            "cmd": "mkdir data",
+            "outs": ["data"],
+        }
+
+    # running by clearing and not clearing stuffs
+    # so as it works even for optimized cases
+    if clear_before_run:
+        remove(data_dir)
+        remove(dvc.cache.local.cache_dir)
+
+    dvc._reset()
+
+    with pytest.raises(OverlappingOutputPathsError):
+        dvc.plots.show()
